pyproject.toml
[project]
name = "fastapi-template"
version = "0.1.0"
description = "FastAPI template with async SQLModel, Postgres, and k8s dev tooling."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "sqlmodel>=0.0.22",
  "asyncpg>=0.29.0",
  "alembic>=1.13.0",
  "pydantic-settings>=2.4.0",
  "prometheus-client>=0.21.0",
  "ecs-logging>=2.2.0",
]

[build-system]
requires = ["hatchling>=1.24.2"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 88
src = ["app"]
show-fixes = true
extend-exclude = ["stubs"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM", "PL", "RUF"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false

[tool.mypy]
python_version = "3.13"
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
ignore_missing_imports = true
mypy_path = "stubs"

[tool.pytest.ini_options]
addopts = "-q"
asyncio_mode = "auto"
testpaths = ["app/tests"]

[dependency-groups]
dev = [
  "pytest>=8.0.0",
  "pytest-asyncio>=0.23.0",
  "pytest-docker>=3.1.1",
  "httpx>=0.27.0",
  "ruff>=0.5.0",
  "mypy>=1.10.0",
  "pre-commit>=3.7.0",
  "aiosqlite>=0.20.0",
]
